[BUGFIX] Function shortcuts were not bound to their component.
When extending ComponentBase and adding to the $funcs[] array so that your functions are easily available in the twig templates, you expect your function to only be called on your component. However, twig stores all functions in an array by name, so two functions with the same name will not resolve properly. To get around this, we're using the twig closure to get the current __SELF__ object and invoking the function on that. Basically, our component functions are no longer going to be global.

One important thing to note is that you should not be able to override a non-component function. I'm not sure how to enforce this yet.